requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.1

# Only required if USE_CLAUDE_CLI=false (SDK mode)
# Uncomment if using SDK mode:
# claude-agent-sdk
# anyio

# Type hints support
typing-extensions>=4.8.0
